+++ crawl-ref/source/sha256.h
@@ -1,7 +1,7 @@
 #ifndef sha256_h
 #define sha256_h
 
-unsigned long sha256_genrand();
+uint32_t sha256_genrand();
 void push_sha256_state();
 void pop_sha256_state();
